Fix exploit for x86_64
Fix large page page table walks. An extra bit was incorrectly being
included in the physical address calculation. On x86_64, 9 bits are
used for the PTE offset instead of 10 (as in non-PAE x86).

Note that the exploit crashes when trying to trigger the
HalDispatchTable overwrite. Microsoft added the HalDispatchTable to
control flow guard, so we end up with a guard_check_icall failure.
Given that we have a full arbitrary read/write, it shouldn't be
difficult to add our payload's page to the CFG bitmap (as long as
VBS is *off*).